summer'25/week2/ideabloom/requirements.txt
fastapi
uvicorn
pydantic
motor
httpx
python-dotenv

